CO2_SCD40/requirements.txt
google-cloud
oauth2-client
adafruit-circuitpython-scd4x
